feat(workspaces): Prompt when deleting workspace

+++ src/apps/main/core/common/workspaces/workspacesService.ts
@@ -64,12 +64,20 @@ export class WorkspacesService implements WorkspacesDataManagerBase {
 
   deleteWorkspace(workspaceID: TWorkspaceID): void {
     if (workspacesDataStore.data.size === 1) return;
-    this.tabManagerCtx.removeTabByWorkspaceId(workspaceID);
-    setWorkspacesDataStore(
-      "order",
-      (prev) => prev.filter((v) => v !== workspaceID),
+    const workspace = this.getRawWorkspace(workspaceID);
+    const result = Services.prompt.confirm(
+      window,
+      i18next.t('workspaces.modal.deleteTitle'),
+      i18next.t('workspaces.modal.deleteText', { name: workspace.name }),
     );
-    this.dataManagerCtx.deleteWorkspace(workspaceID);
+    if (result) {
+      this.tabManagerCtx.removeTabByWorkspaceId(workspaceID);
+      setWorkspacesDataStore(
+          "order",
+          (prev) => prev.filter((v) => v !== workspaceID),
+      );
+      this.dataManagerCtx.deleteWorkspace(workspaceID);
+    }
   }
 
   /**
